Move common entity conversion logic into the abstract base
