+++ server/src/main/java/io/spine/server/entity/EntityStorageConverter.java
@@ -21,12 +21,19 @@
 package io.spine.server.entity;
 
 import com.google.common.base.Converter;
+import com.google.protobuf.Any;
 import com.google.protobuf.FieldMask;
 import com.google.protobuf.Message;
+import io.spine.base.Identifier;
+import io.spine.type.TypeUrl;
 
+import javax.annotation.Nullable;
 import java.io.Serializable;
+import java.util.Objects;
 
 import static com.google.common.base.Preconditions.checkNotNull;
+import static io.spine.protobuf.AnyPacker.pack;
+import static io.spine.protobuf.AnyPacker.unpack;
 
 /**
  * An abstract base for converters of entities into {@link EntityRecord}.
@@ -37,15 +44,36 @@ public abstract class EntityStorageConverter<I, E extends Entity<I, S>, S extend
         extends Converter<E, EntityRecord> implements Serializable {
 
     private static final long serialVersionUID = 0L;
+    private final TypeUrl entityStateType;
+    private final EntityFactory<I, E> entityFactory;
     private final FieldMask fieldMask;
 
-    protected EntityStorageConverter(FieldMask fieldMask) {
+    protected EntityStorageConverter(TypeUrl entityStateType,
+                                     EntityFactory<I, E> factory,
+                                     FieldMask fieldMask) {
         super();
         this.fieldMask = checkNotNull(fieldMask);
+        this.entityFactory = factory;
+        this.entityStateType = entityStateType;
     }
 
     /**
-     * Obtains the field mask used by this converter.
+     * Obtains the type URL of the state of entities which this converter builds.
+     */
+    protected TypeUrl getEntityStateType() {
+        return entityStateType;
+    }
+
+    /**
+     * Obtains the entity factory used by the converter.
+     */
+    protected EntityFactory<I, E> getEntityFactory() {
+        return entityFactory;
+    }
+
+    /**
+     * Obtains the field mask used by this converter to trim the state of entities before the state
+     * is {@linkplain #injectState(Entity, Message, EntityRecord) injected} into entities.
      */
     protected FieldMask getFieldMask() {
         return this.fieldMask;
@@ -55,4 +83,71 @@ public abstract class EntityStorageConverter<I, E extends Entity<I, S>, S extend
      * Creates a copy of this converter modified with the passed filed mask.
      */
     public abstract EntityStorageConverter<I, E, S> withFieldMask(FieldMask fieldMask);
+
+    @Override
+    protected EntityRecord doForward(E entity) {
+        final Any entityId = Identifier.pack(entity.getId());
+        final Any stateAny = pack(entity.getState());
+        final EntityRecord.Builder builder = EntityRecord.newBuilder()
+                                                         .setEntityId(entityId)
+                                                         .setState(stateAny);
+        updateBuilder(builder, entity);
+        return builder.build();
+    }
+
+    /**
+     * Derived classes may override to additionally tune the passed entity builder.
+     *
+     * <p>Default implementation does nothing.
+     *
+     * @param builder the entity builder to update
+     * @param entity  the entity which data is passed to the {@link EntityRecord} we are building
+     */
+    @SuppressWarnings("NoopMethodInAbstractClass") // Avoid forcing empty implementations.
+    protected void updateBuilder(EntityRecord.Builder builder, E entity) {
+        // Do nothing.
+    }
+
+    @Override
+    @SuppressWarnings("unchecked")
+    protected E doBackward(EntityRecord entityRecord) {
+        final Message unpacked = unpack(entityRecord.getState());
+        final S state = (S) FieldMasks.applyMask(getFieldMask(), unpacked, entityStateType);
+
+        final I id = (I) Identifier.unpack(entityRecord.getEntityId());
+        final E entity = entityFactory.create(id);
+
+        if (entity != null) {
+            injectState(entity, state, entityRecord);
+        }
+        return entity;
+    }
+
+    /**
+     * Derived classes must implement providing state injection into the passed entity.
+     *
+     * @param entity       the entity into which inject the state
+     * @param state        the state message extracted from the record
+     * @param entityRecord the record which may contain additional properties for the entity
+     */
+    protected abstract void injectState(E entity, S state, EntityRecord entityRecord);
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(entityFactory, entityStateType, fieldMask);
+    }
+
+    @Override
+    public boolean equals(@Nullable Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null || getClass() != obj.getClass()) {
+            return false;
+        }
+        final EntityStorageConverter other = (EntityStorageConverter) obj;
+        return Objects.equals(this.entityStateType, other.entityStateType)
+                && Objects.equals(this.entityFactory, other.entityFactory)
+                && Objects.equals(this.fieldMask, other.fieldMask);
+    }
 }
